feat: reorganize folders around to give room to grow.
