+++ src/app.rs
@@ -1,10 +1,8 @@
-use axum::Router;
 use tokio::sync::broadcast::{self, Receiver, Sender};
 use tracing::info;
-use utoipa::OpenApi;
-use utoipa_swagger_ui::SwaggerUi;
 
 use crate::{
+    api::router,
     error::NetherilErr,
     logging::{Logging, LoggingOptions},
 };
@@ -94,254 +92,8 @@ async fn register_signals(broadcast: Sender<Broadcast>) -> Result<(), Box<dyn st
     }
 }
 
-fn router() -> Router {
-    //.merge(self.swagger_ui())
-    Router::new().merge(swagger_ui()).nest(
-        "/api/",
-        root::router().nest("/operations/", operations::router()),
-    )
-}
-
-pub fn swagger_ui() -> SwaggerUi {
-    #[derive(OpenApi)]
-    #[openapi(
-	nest(
-	    (path = "/api", api = root::ApiDoc),
-	    (path = "/api/operations/", api = operations::ApiDoc),
-	)
-    )]
-    struct ApiDoc;
-
-    let doc = ApiDoc::openapi();
-
-    SwaggerUi::new("/swagger-ui").url("/api-docs/openapi.json", doc)
-}
-
 impl Default for App {
     fn default() -> Self {
         Self::new()
-    }
-}
-
-#[allow(unused)]
-struct OperationService {}
-
-#[allow(unused)]
-impl OperationService {
-    pub fn new() -> Self {
-        Self {}
-    }
-
-    pub fn find(&self, id: &str) {
-        println!("find: {}", id);
-    }
-}
-
-mod operations {
-    use axum::{extract::Path, http::StatusCode, routing::get, Json, Router};
-    use serde::{Deserialize, Serialize};
-    use utoipa::{OpenApi, ToSchema};
-
-    #[derive(OpenApi)]
-    #[openapi(paths(show))]
-    pub struct ApiDoc;
-
-    pub fn router() -> Router {
-        Router::new().route("/{id}", get(show))
-    }
-
-    #[derive(Debug, Deserialize)]
-    struct ShowPath {
-        id: String,
-    }
-
-    #[derive(Debug, Serialize, ToSchema)]
-    #[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-    enum Status {
-        // Completed,
-        // Error,
-        // Queued,
-        InProgress,
-    }
-
-    #[derive(Debug, Serialize, ToSchema)]
-    struct ShowResponse {
-        operation_id: String,
-        status: Status,
-    }
-
-    #[utoipa::path(
-	get,
-	path = "/operations/:id",
-	responses(
-	    (status = OK, description = "Successfully retrieve the specified operation", body = ShowResponse)
-	)
-    )]
-    async fn show(Path(ShowPath { id }): Path<ShowPath>) -> (StatusCode, Json<ShowResponse>) {
-        (
-            StatusCode::OK,
-            Json(ShowResponse {
-                operation_id: id,
-                status: Status::InProgress,
-            }),
-        )
-    }
-}
-
-mod root {
-    use axum::{http::StatusCode, routing::get, Json, Router};
-    use serde::Serialize;
-    use utoipa::{OpenApi, ToSchema};
-
-    use crate::version::{self, Build};
-
-    #[derive(OpenApi)]
-    #[openapi(paths(index))]
-    pub struct ApiDoc;
-
-    pub fn router() -> Router {
-        Router::new().route("/", get(index))
-    }
-
-    #[derive(Serialize, ToSchema)]
-    pub struct RootResponse {
-        message: &'static str,
-        build: Build,
-    }
-
-    impl Default for RootResponse {
-        fn default() -> Self {
-            RootResponse {
-                message: "Hello from Netheril",
-                build: version::BUILD,
-            }
-        }
-    }
-
-    #[utoipa::path(
-	get,
-	path = "/",
-	responses(
-	    (status = OK, body = RootResponse)
-	)
-    )]
-    async fn index() -> (StatusCode, Json<RootResponse>) {
-        (StatusCode::OK, Json(RootResponse::default()))
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use crate::version::BUILD;
-    use reqwest::RequestBuilder;
-    use serde::Deserialize;
-    use std::net::SocketAddr;
-    use tokio::task::JoinHandle;
-
-    use super::*;
-
-    #[derive(Debug)]
-    struct Client {
-        addr: SocketAddr,
-        client: reqwest::Client,
-    }
-
-    impl Client {
-        fn new(addr: SocketAddr) -> Result<Self, Box<dyn std::error::Error>> {
-            let client = reqwest::Client::builder().build()?;
-            Ok(Self { addr, client })
-        }
-
-        fn get<R: Into<RelativeUrl>>(&self, path: R) -> RequestBuilder {
-            let url = self.base_url(path.into());
-            self.client.get(url)
-        }
-
-        fn base_url(&self, path: RelativeUrl) -> String {
-            format!("http://{}:{}{}", self.addr.ip(), self.addr.port(), path)
-        }
-    }
-
-    #[derive(Debug)]
-    struct RelativeUrl(String);
-
-    impl std::fmt::Display for RelativeUrl {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            write!(f, "{}", self.0)
-        }
-    }
-
-    impl From<&str> for RelativeUrl {
-        fn from(value: &str) -> Self {
-            const HTTP_PREFIX: &str = "http://";
-            const HTTPS_PREFIX: &str = "https://";
-
-            let candidate = value.to_lowercase();
-
-            if candidate.starts_with(HTTP_PREFIX) && candidate.starts_with(HTTPS_PREFIX) {
-                panic!("bad relative url: `{}`", value)
-            }
-
-            RelativeUrl(value.to_string())
-        }
-    }
-
-    #[derive(Debug)]
-    struct TestServer {
-        server_handle: JoinHandle<()>,
-    }
-
-    impl TestServer {
-        async fn new(router: Router) -> Result<(Self, Client), Box<dyn std::error::Error>> {
-            const ANY_LOCAL_PORT: &str = "0.0.0.0:0";
-
-            let listener = tokio::net::TcpListener::bind(ANY_LOCAL_PORT).await?;
-            let addr = listener.local_addr()?;
-
-            let server_handle = tokio::spawn(async move {
-                axum::serve(listener, router).await.unwrap();
-            });
-
-            Ok((Self { server_handle }, Client::new(addr)?))
-        }
-    }
-
-    impl Drop for TestServer {
-        fn drop(&mut self) {
-            self.server_handle.abort()
-        }
-    }
-
-    #[tokio::test]
-    async fn it_should_return_the_build_information() {
-        #[derive(Deserialize)]
-        struct BuildResponse {
-            version: String,
-            git_sha: String,
-            build_date: String,
-        }
-
-        #[derive(Deserialize)]
-        struct Response {
-            message: String,
-            build: BuildResponse,
-        }
-
-        let (_server, client) = TestServer::new(router()).await.unwrap();
-
-        let response: Response = client
-            .get("/api/")
-            .send()
-            .await
-            .unwrap()
-            .json()
-            .await
-            .unwrap();
-
-        assert_eq!(response.message, "Hello from Netheril");
-
-        assert_eq!(response.build.version, BUILD.version);
-        assert_eq!(response.build.git_sha, BUILD.git_sha);
-        assert_eq!(response.build.build_date, BUILD.build_date);
     }
 }
